pyproject.toml
[project]
name = "BayAKM"
version = "1.0.0"
description = "A tool for Bayesian optimization of chemical reactions."
authors = [
    { name = "Niklas Abe", email = "niklas.abe@hhu.de" }
]
dependencies = [
    "customtkinter",
    "baybe[chem,insights]<=0.14.0",
    "pyyaml",
    "pandas",
    "numpy"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"